refactor: move query tests to `src/queries/__tests__`
