+++ src/queries/__tests__/a11yAPI.test.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
 import { TouchableOpacity, Text } from 'react-native';
-import { render } from '../index';
+import { render } from '../..';
 
 const BUTTON_LABEL = 'cool button';
 const BUTTON_HINT = 'click this button';
